requirements.txt
Flask
requests
python-dotenv
python-dotenv
nltk 
fuzzywuzzy
scikit-learn 
python-Levenshtein
sentence-transformers
numpy
